Revert to Repo, for minor version release.
- Restored verbose mode for URL parsing
- Converted back to repo naming
- Added parameter names
- Removed redundant parameter casting
- Added commented events request, ready to be enabled oce Repo is converted to Project
