+++ TanukiKitTests/Fixtures/Projects.json
@@ -1,75 +1,75 @@
 [
-	{
-		"id": 3,
-		"description": null,
-		"default_branch": "master",
-		"public": false,
-		"visibility_level": 0,
-		"ssh_url_to_repo": "git@example.com:diaspora/diaspora-project-site.git",
-		"http_url_to_repo": "http://example.com/diaspora/diaspora-project-site.git",
-		"web_url": "http://example.com/diaspora/diaspora-project-site",
-		"tag_list": [
-			"example",
-			"disapora project"
-		],
-		"owner": {
-			"id": 3,
-			"name": "Diaspora",
-			"created_at": "2013-09-30T13:46:02Z"
-		},
-		"name": "Diaspora Project Site",
-		"name_with_namespace": "Diaspora / Diaspora Project Site",
-		"path": "diaspora-project-site",
-		"path_with_namespace": "diaspora/diaspora-project-site",
-		"issues_enabled": true,
-		"open_issues_count": 1,
-		"merge_requests_enabled": true,
-		"builds_enabled": true,
-		"wiki_enabled": true,
-		"snippets_enabled": false,
-		"container_registry_enabled": false,
-		"created_at": "2013-09-30T13:46:02Z",
-		"last_activity_at": "2013-09-30T13:46:02Z",
-		"creator_id": 3,
-		"namespace": {
-			"created_at": "2013-09-30T13:46:02Z",
-			"description": "",
-			"id": 3,
-			"name": "Diaspora",
-			"owner_id": 1,
-			"path": "diaspora",
-			"updated_at": "2013-09-30T13:46:02Z"
-		},
-		"permissions": {
-			"project_access": {
-				"access_level": 10,
-				"notification_level": 3
-			},
-			"group_access": {
-				"access_level": 50,
-				"notification_level": 3
-			}
-		},
-		"archived": false,
-		"avatar_url": "http://example.com/uploads/project/avatar/3/uploads/avatar.png",
-		"shared_runners_enabled": true,
-		"forks_count": 0,
-		"star_count": 0,
-		"runners_token": "b8bc4a7a29eb76ea83cf79e4908c2b",
-		"public_builds": true,
-		"shared_with_groups": [
-			{
-				"group_id": 4,
-				"group_name": "Twitter",
-				"group_access_level": 30
-			},
-			{
-				"group_id": 3,
-				"group_name": "Gitlab Org",
-				"group_access_level": 10
-			}
-		],
-		"only_allow_merge_if_build_succeeds": false,
-		"request_access_enabled": false
-	}
+  {
+    "id": 3,
+    "description": null,
+    "default_branch": "master",
+    "public": false,
+    "visibility_level": 0,
+    "ssh_url_to_repo": "git@example.com:diaspora/diaspora-project-site.git",
+    "http_url_to_repo": "http://example.com/diaspora/diaspora-project-site.git",
+    "web_url": "http://example.com/diaspora/diaspora-project-site",
+    "tag_list": [
+      "example",
+      "disapora project"
+    ],
+    "owner": {
+      "id": 3,
+      "name": "Diaspora",
+      "created_at": "2013-09-30T13:46:02Z"
+    },
+    "name": "Diaspora Project Site",
+    "name_with_namespace": "Diaspora / Diaspora Project Site",
+    "path": "diaspora-project-site",
+    "path_with_namespace": "diaspora/diaspora-project-site",
+    "issues_enabled": true,
+    "open_issues_count": 1,
+    "merge_requests_enabled": true,
+    "builds_enabled": true,
+    "wiki_enabled": true,
+    "snippets_enabled": false,
+    "container_registry_enabled": false,
+    "created_at": "2013-09-30T13:46:02Z",
+    "last_activity_at": "2013-09-30T13:46:02Z",
+    "creator_id": 3,
+    "namespace": {
+      "created_at": "2013-09-30T13:46:02Z",
+      "description": "",
+      "id": 3,
+      "name": "Diaspora",
+      "owner_id": 1,
+      "path": "diaspora",
+      "updated_at": "2013-09-30T13:46:02Z"
+    },
+    "permissions": {
+      "project_access": {
+        "access_level": 10,
+        "notification_level": 3
+      },
+      "group_access": {
+        "access_level": 50,
+        "notification_level": 3
+      }
+    },
+    "archived": false,
+    "avatar_url": "http://example.com/uploads/project/avatar/3/uploads/avatar.png",
+    "shared_runners_enabled": true,
+    "forks_count": 0,
+    "star_count": 0,
+    "runners_token": "b8bc4a7a29eb76ea83cf79e4908c2b",
+    "public_builds": true,
+    "shared_with_groups": [
+      {
+        "group_id": 4,
+        "group_name": "Twitter",
+        "group_access_level": 30
+      },
+      {
+        "group_id": 3,
+        "group_name": "Gitlab Org",
+        "group_access_level": 10
+      }
+    ],
+    "only_allow_merge_if_build_succeeds": false,
+    "request_access_enabled": false
+  }
 ]
