+++ TanukiKitTests/Fixtures/User.json
@@ -1,28 +1,28 @@
 {
-	"name": "Test McTestface",
-	"username": "testmctestface",
-	"id": 7,
-	"state": "active",
-	"avatar_url": "https://code.tiferrei.com/uploads/user/avatar/7/avatar.png",
-	"web_url": "https://code.tiferrei.com/u/testmctestface",
-	"created_at": "2016-05-03T15:05:46.391Z",
-	"is_admin": false,
-	"bio": "I'm a simple test user that Tiago created to test the GitLab API.",
-	"location": "World Wide Web",
-	"skype": "testMcTestface",
-	"linkedin": "testMcTestface",
-	"twitter": "@testMcTestface",
-	"website_url": "https://testmctestface.example.com",
-	"last_sign_in_at": "2016-05-03T15:06:21.305Z",
-	"confirmed_at": "2016-05-03T15:05:46.183Z",
-	"email": "EMAIL",
-	"theme_id": 2,
-	"color_scheme_id": 1,
-	"projects_limit": 10,
-	"current_sign_in_at": "2016-06-26T15:29:16.606Z",
-	"identities": [],
-	"can_create_group": false,
-	"can_create_project": true,
-	"two_factor_enabled": false,
-	"external": false
+  "name": "Test McTestface",
+  "username": "testmctestface",
+  "id": 7,
+  "state": "active",
+  "avatar_url": "https://code.tiferrei.com/uploads/user/avatar/7/avatar.png",
+  "web_url": "https://code.tiferrei.com/u/testmctestface",
+  "created_at": "2016-05-03T15:05:46.391Z",
+  "is_admin": false,
+  "bio": "I'm a simple test user that Tiago created to test the GitLab API.",
+  "location": "World Wide Web",
+  "skype": "testMcTestface",
+  "linkedin": "testMcTestface",
+  "twitter": "@testMcTestface",
+  "website_url": "https://testmctestface.example.com",
+  "last_sign_in_at": "2016-05-03T15:06:21.305Z",
+  "confirmed_at": "2016-05-03T15:05:46.183Z",
+  "email": "EMAIL",
+  "theme_id": 2,
+  "color_scheme_id": 1,
+  "projects_limit": 10,
+  "current_sign_in_at": "2016-06-26T15:29:16.606Z",
+  "identities": [],
+  "can_create_group": false,
+  "can_create_project": true,
+  "two_factor_enabled": false,
+  "external": false
 }
